prevent "config.log_level" DEPRECATION WARNING
DEPRECATION WARNING: You did not specify a `log_level` in `production.rb`.
Currently, the default value for `log_level` is `:info` for the production environment
and `:debug` in all other environments.
In Rails 5 the default value will be unified to `:debug` across all environments.
To preserve the current setting, add the following line to your `production.rb`:

   config.log_level = :info