[#20288] Update styles to match CodeRay 1.1.0 (preserving changes of r10132).
This commit updates the CSS styles to match with CodeRay 1.1.0,
while it preserves the custom changes applied in r10132.

The CSS styles were still based on CodeRay 1.0.6 (included since
Redmine 1.4.0) with the custom changes from r10132 (included since
Redmine 2.1.0). Redmine 2.3.2 till 2.3.4 came with CodeRay 1.0.9,
an upgrade that didn't needed changes in the CSS styles. Starting
with 2.4.0 Redmine comes with CodeRay 1.1.0, a minor upgrade
that came with new/changed token_kinds and lots of changes in the
alpha stylesheet, that in turn is used as a base for Redmine's own
CodeRay CSS styles. As such, this upgrade needed CSS stylesheet
changes like done before in r7618 and r7623 (for 1.0.0 upgrade)
and r9389 (for 1.0.6 upgrade). But these changes, plus an update
of the Redmine core documentation that is shipped along the core
(wiki_syntax_detailed_[markdown|textile].html), aren't integrated
up untill today.

Contributed by Mischa The Evil.