Fix HTML structure of query_form on calendar/gantt views (#26125).
This unifies the HTML structure of the query form on the calendars/show
and gantts/show views to be the same as it is currently for the issues
and spent time views, which were themself unified into the new
queries/_query_form partial in r15649.

Patch by Mischa The Evil.